Split resource.validate tests
