+++ tests/resource/actions/validate/test_general.py
@@ -168,738 +168,6 @@ def test_validate_source_pathlib_path_table():
     assert report.valid
 
 
-# Scheme
-
-
-def test_validate_scheme():
-    report = validate("data/table.csv", scheme="file")
-    assert report.valid
-
-
-def test_validate_scheme_invalid():
-    report = validate("bad://data/table.csv")
-    assert report.flatten(["code", "note"]) == [
-        ["scheme-error", 'cannot create loader "bad". Try installing "frictionless-bad"'],
-    ]
-
-
-# Format
-
-
-def test_validate_format():
-    report = validate("data/table.csv", format="csv")
-    assert report.valid
-
-
-def test_validate_format_non_tabular():
-    report = validate("data/table.bad")
-    assert report.valid
-
-
-# Encoding
-
-
-def test_validate_encoding():
-    report = validate("data/table.csv", encoding="utf-8")
-    assert report.valid
-
-
-def test_validate_encoding_invalid():
-    report = validate("data/latin1.csv", encoding="utf-8")
-    assert not report.valid
-    if IS_UNIX:
-        assert report.flatten(["code", "note"]) == [
-            [
-                "encoding-error",
-                "'utf-8' codec can't decode byte 0xa9 in position 20: invalid start byte",
-            ],
-        ]
-
-
-# Compression
-
-
-def test_validate_compression():
-    report = validate("data/table.csv.zip")
-    assert report.valid
-
-
-def test_validate_compression_explicit():
-    report = validate("data/table.csv.zip", compression="zip")
-    assert report.valid
-
-
-def test_validate_compression_invalid():
-    report = validate("data/table.csv.zip", compression="bad")
-    assert report.flatten(["code", "note"]) == [
-        ["compression-error", 'compression "bad" is not supported'],
-    ]
-
-
-# Dialect
-
-
-def test_validate_dialect_delimiter():
-    report = validate("data/delimiter.csv", dialect={"delimiter": ";"})
-    assert report.valid
-    assert report.task.resource.stats["rows"] == 2
-
-
-# Layout
-
-
-def test_validate_layout_none():
-    layout = Layout(header=False)
-    report = validate("data/without-headers.csv", layout=layout)
-    assert report.valid
-    assert report.task.resource.stats["rows"] == 3
-    assert report.task.resource.layout.header is False
-    assert report.task.resource.labels == []
-    assert report.task.resource.header == ["field1", "field2"]
-
-
-def test_validate_layout_none_extra_cell():
-    layout = Layout(header=False)
-    report = validate("data/without-headers-extra.csv", layout=layout)
-    assert report.task.resource.stats["rows"] == 3
-    assert report.task.resource.layout.header is False
-    assert report.task.resource.labels == []
-    assert report.task.resource.header == ["field1", "field2"]
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [3, 3, "extra-cell"],
-    ]
-
-
-def test_validate_layout_number():
-    layout = Layout(header_rows=[2])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["11", "12", "13", "14"]
-    assert report.valid
-
-
-def test_validate_layout_list_of_numbers():
-    layout = Layout(header_rows=[2, 3, 4])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["11 21 31", "12 22 32", "13 23 33", "14 24 34"]
-    assert report.valid
-
-
-def test_validate_layout_list_of_numbers_and_headers_join():
-    layout = Layout(header_rows=[2, 3, 4], header_join=".")
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["11.21.31", "12.22.32", "13.23.33", "14.24.34"]
-    assert report.valid
-
-
-def test_validate_layout_pick_fields():
-    layout = Layout(pick_fields=[2, "f3"])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f2", "f3"]
-    assert report.task.resource.stats["rows"] == 4
-    assert report.task.valid
-
-
-def test_validate_layout_pick_fields_regex():
-    layout = Layout(pick_fields=["<regex>f[23]"])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f2", "f3"]
-    assert report.task.resource.stats["rows"] == 4
-    assert report.task.valid
-
-
-def test_validate_layout_skip_fields():
-    layout = Layout(skip_fields=[1, "f4"])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f2", "f3"]
-    assert report.task.resource.stats["rows"] == 4
-    assert report.task.valid
-
-
-def test_validate_layout_skip_fields_regex():
-    layout = Layout(skip_fields=["<regex>f[14]"])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f2", "f3"]
-    assert report.task.resource.stats["rows"] == 4
-    assert report.task.valid
-
-
-def test_validate_layout_limit_fields():
-    layout = Layout(limit_fields=1)
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f1"]
-    assert report.task.resource.stats["rows"] == 4
-    assert report.task.valid
-
-
-def test_validate_layout_offset_fields():
-    layout = Layout(offset_fields=3)
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f4"]
-    assert report.task.resource.stats["rows"] == 4
-    assert report.task.valid
-
-
-def test_validate_layout_limit_and_offset_fields():
-    layout = Layout(limit_fields=2, offset_fields=1)
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f2", "f3"]
-    assert report.task.resource.stats["rows"] == 4
-    assert report.task.valid
-
-
-def test_validate_layout_pick_rows():
-    layout = Layout(pick_rows=[1, 3, "31"])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f1", "f2", "f3", "f4"]
-    assert report.task.resource.stats["rows"] == 2
-    assert report.task.valid
-
-
-def test_validate_layout_pick_rows_regex():
-    layout = Layout(pick_rows=["<regex>[f23]1"])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f1", "f2", "f3", "f4"]
-    assert report.task.resource.stats["rows"] == 2
-    assert report.task.valid
-
-
-def test_validate_layout_skip_rows():
-    layout = Layout(skip_rows=[2, "41"])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f1", "f2", "f3", "f4"]
-    assert report.task.resource.stats["rows"] == 2
-    assert report.task.valid
-
-
-def test_validate_layout_skip_rows_regex():
-    layout = Layout(skip_rows=["<regex>[14]1"])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f1", "f2", "f3", "f4"]
-    assert report.task.resource.stats["rows"] == 2
-    assert report.task.valid
-
-
-def test_validate_layout_skip_rows_blank():
-    layout = Layout(skip_rows=["<blank>"])
-    report = validate("data/blank-rows.csv", layout=layout)
-    assert report.task.resource.header == ["id", "name", "age"]
-    assert report.task.resource.stats["rows"] == 2
-    assert report.task.valid
-
-
-def test_validate_layout_pick_rows_and_fields():
-    layout = Layout(pick_rows=[1, 3, "31"], pick_fields=[2, "f3"])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f2", "f3"]
-    assert report.task.resource.stats["rows"] == 2
-    assert report.task.valid
-
-
-def test_validate_layout_skip_rows_and_fields():
-    layout = Layout(skip_rows=[2, "41"], skip_fields=[1, "f4"])
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f2", "f3"]
-    assert report.task.resource.stats["rows"] == 2
-    assert report.task.valid
-
-
-def test_validate_layout_limit_rows():
-    layout = Layout(limit_rows=1)
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f1", "f2", "f3", "f4"]
-    assert report.task.resource.stats["rows"] == 1
-    assert report.task.valid
-
-
-def test_validate_layout_offset_rows():
-    layout = Layout(offset_rows=3)
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f1", "f2", "f3", "f4"]
-    assert report.task.resource.stats["rows"] == 1
-    assert report.task.valid
-
-
-def test_validate_layout_limit_and_offset_rows():
-    layout = Layout(limit_rows=2, offset_rows=1)
-    report = validate("data/matrix.csv", layout=layout)
-    assert report.task.resource.header == ["f1", "f2", "f3", "f4"]
-    assert report.task.resource.stats["rows"] == 2
-    assert report.task.valid
-
-
-def test_validate_layout_invalid_limit_rows():
-    layout = Layout(limit_rows=2)
-    report = validate("data/invalid.csv", layout=layout)
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [None, 3, "blank-label"],
-        [None, 4, "duplicate-label"],
-        [2, 3, "missing-cell"],
-        [2, 4, "missing-cell"],
-        [3, 3, "missing-cell"],
-        [3, 4, "missing-cell"],
-    ]
-
-
-def test_validate_layout_structure_errors_with_limit_rows():
-    layout = Layout(limit_rows=3)
-    report = validate("data/structure-errors.csv", layout=layout)
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [4, None, "blank-row"],
-    ]
-
-
-# Schema
-
-
-def test_validate_schema_invalid():
-    source = [["name", "age"], ["Alex", "33"]]
-    schema = {"fields": [{"name": "name"}, {"name": "age", "type": "bad"}]}
-    report = validate(source, schema=schema)
-    assert report.flatten(["code", "note"]) == [
-        [
-            "field-error",
-            "\"{'name': 'age', 'type': 'bad'} is not valid under any of the given schemas\" at \"\" in metadata and at \"anyOf\" in profile",
-        ],
-    ]
-
-
-def test_validate_schema_invalid_json():
-    report = validate("data/table.csv", schema="data/invalid.json")
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [None, None, "schema-error"],
-    ]
-
-
-def test_validate_schema_extra_headers_and_cells():
-    schema = {"fields": [{"name": "id", "type": "integer"}]}
-    report = validate("data/table.csv", schema=schema)
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [None, 2, "extra-label"],
-        [2, 2, "extra-cell"],
-        [3, 2, "extra-cell"],
-    ]
-
-
-def test_validate_schema_multiple_errors():
-    source = "data/schema-errors.csv"
-    schema = "data/schema-valid.json"
-    report = validate(source, schema=schema, pick_errors=["#row"], limit_errors=3)
-    assert report.task.partial
-    assert report.task.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [4, 1, "type-error"],
-        [4, 2, "constraint-error"],
-        [4, 3, "constraint-error"],
-    ]
-
-
-def test_validate_schema_min_length_constraint():
-    source = [["row", "word"], [2, "a"], [3, "ab"], [4, "abc"], [5, "abcd"], [6]]
-    schema = {
-        "fields": [
-            {"name": "row", "type": "integer"},
-            {"name": "word", "type": "string", "constraints": {"minLength": 2}},
-        ]
-    }
-    report = validate(source, schema=schema, pick_errors=["constraint-error"])
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [2, 2, "constraint-error"],
-    ]
-
-
-def test_validate_schema_max_length_constraint():
-    source = [["row", "word"], [2, "a"], [3, "ab"], [4, "abc"], [5, "abcd"], [6]]
-    schema = {
-        "fields": [
-            {"name": "row", "type": "integer"},
-            {"name": "word", "type": "string", "constraints": {"maxLength": 2}},
-        ]
-    }
-    report = validate(source, schema=schema, pick_errors=["constraint-error"])
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [4, 2, "constraint-error"],
-        [5, 2, "constraint-error"],
-    ]
-
-
-def test_validate_schema_minimum_constraint():
-    source = [["row", "score"], [2, 1], [3, 2], [4, 3], [5, 4], [6]]
-    schema = {
-        "fields": [
-            {"name": "row", "type": "integer"},
-            {"name": "score", "type": "integer", "constraints": {"minimum": 2}},
-        ]
-    }
-    report = validate(source, schema=schema, pick_errors=["constraint-error"])
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [2, 2, "constraint-error"],
-    ]
-
-
-def test_validate_schema_maximum_constraint():
-    source = [["row", "score"], [2, 1], [3, 2], [4, 3], [5, 4], [6]]
-    schema = {
-        "fields": [
-            {"name": "row", "type": "integer"},
-            {"name": "score", "type": "integer", "constraints": {"maximum": 2}},
-        ]
-    }
-    report = validate(source, schema=schema, pick_errors=["constraint-error"])
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [4, 2, "constraint-error"],
-        [5, 2, "constraint-error"],
-    ]
-
-
-def test_validate_schema_foreign_key_error_self_referencing():
-    source = {
-        "path": "data/nested.csv",
-        "schema": {
-            "fields": [
-                {"name": "id", "type": "integer"},
-                {"name": "cat", "type": "integer"},
-                {"name": "name", "type": "string"},
-            ],
-            "foreignKeys": [
-                {"fields": "cat", "reference": {"resource": "", "fields": "id"}}
-            ],
-        },
-    }
-    report = validate(source)
-    assert report.valid
-
-
-def test_validate_schema_foreign_key_error_self_referencing_invalid():
-    source = {
-        "path": "data/nested-invalid.csv",
-        "schema": {
-            "fields": [
-                {"name": "id", "type": "integer"},
-                {"name": "cat", "type": "integer"},
-                {"name": "name", "type": "string"},
-            ],
-            "foreignKeys": [
-                {"fields": "cat", "reference": {"resource": "", "fields": "id"}}
-            ],
-        },
-    }
-    report = validate(source)
-    assert report.flatten(["rowPosition", "fieldPosition", "code", "cells"]) == [
-        [6, None, "foreign-key-error", ["5", "6", "Rome"]],
-    ]
-
-
-def test_validate_schema_unique_error():
-    report = validate(
-        "data/unique-field.csv",
-        schema="data/unique-field.json",
-        pick_errors=["unique-error"],
-    )
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [10, 1, "unique-error"],
-    ]
-
-
-def test_validate_schema_unique_error_and_type_error():
-    source = [
-        ["id", "unique_number"],
-        ["a1", 100],
-        ["a2", "bad"],
-        ["a3", 100],
-        ["a4", 0],
-        ["a5", 0],
-    ]
-    schema = {
-        "fields": [
-            {"name": "id"},
-            {"name": "unique_number", "type": "number", "constraints": {"unique": True}},
-        ]
-    }
-    report = validate(source, schema=schema)
-    assert report.flatten(["rowPosition", "fieldPosition", "code", "cells"]) == [
-        [3, 2, "type-error", ["a2", "bad"]],
-        [4, 2, "unique-error", ["a3", "100"]],
-        [6, 2, "unique-error", ["a5", "0"]],
-    ]
-
-
-def test_validate_schema_primary_key_error():
-    report = validate(
-        "data/unique-field.csv",
-        schema="data/unique-field.json",
-        pick_errors=["primary-key-error"],
-    )
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [10, None, "primary-key-error"],
-    ]
-
-
-def test_validate_schema_primary_key_and_unique_error():
-    report = validate(
-        "data/unique-field.csv",
-        schema="data/unique-field.json",
-    )
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [10, 1, "unique-error"],
-        [10, None, "primary-key-error"],
-    ]
-
-
-def test_validate_schema_primary_key_error_composite():
-    source = [
-        ["id", "name"],
-        [1, "Alex"],
-        [1, "John"],
-        ["", "Paul"],
-        [1, "John"],
-        ["", None],
-    ]
-    schema = {
-        "fields": [
-            {"name": "id", "type": "integer"},
-            {"name": "name", "type": "string"},
-        ],
-        "primaryKey": ["id", "name"],
-    }
-    report = validate(source, schema=schema)
-    assert report.flatten(["rowPosition", "fieldPosition", "code"]) == [
-        [5, None, "primary-key-error"],
-        [6, None, "blank-row"],
-        [6, None, "primary-key-error"],
-    ]
-
-
-# Stats
-
-
-def test_validate_stats_hash():
-    hash = "6c2c61dd9b0e9c6876139a449ed87933"
-    report = validate("data/table.csv", stats={"hash": hash})
-    if IS_UNIX:
-        assert report.task.valid
-
-
-def test_validate_stats_hash_invalid():
-    hash = "6c2c61dd9b0e9c6876139a449ed87933"
-    report = validate("data/table.csv", stats={"hash": "bad"})
-    if IS_UNIX:
-        assert report.flatten(["code", "note"]) == [
-            ["hash-count-error", 'expected md5 is "bad" and actual is "%s"' % hash],
-        ]
-
-
-def test_validate_stats_hash_md5():
-    hash = "6c2c61dd9b0e9c6876139a449ed87933"
-    report = validate("data/table.csv", stats={"hash": hash})
-    if IS_UNIX:
-        assert report.task.valid
-
-
-def test_validate_stats_hash_md5_invalid():
-    hash = "6c2c61dd9b0e9c6876139a449ed87933"
-    report = validate("data/table.csv", stats={"hash": "bad"})
-    if IS_UNIX:
-        assert report.flatten(["code", "note"]) == [
-            ["hash-count-error", 'expected md5 is "bad" and actual is "%s"' % hash],
-        ]
-
-
-def test_validate_stats_hash_sha1():
-    hash = "db6ea2f8ff72a9e13e1d70c28ed1c6b42af3bb0e"
-    report = validate("data/table.csv", hashing="sha1", stats={"hash": hash})
-    if IS_UNIX:
-        assert report.task.valid
-
-
-def test_validate_stats_hash_sha1_invalid():
-    hash = "db6ea2f8ff72a9e13e1d70c28ed1c6b42af3bb0e"
-    report = validate("data/table.csv", hashing="sha1", stats={"hash": "bad"})
-    if IS_UNIX:
-        assert report.flatten(["code", "note"]) == [
-            ["hash-count-error", 'expected sha1 is "bad" and actual is "%s"' % hash],
-        ]
-
-
-def test_validate_stats_hash_sha256():
-    hash = "a1fd6c5ff3494f697874deeb07f69f8667e903dd94a7bc062dd57550cea26da8"
-    report = validate("data/table.csv", hashing="sha256", stats={"hash": hash})
-    if IS_UNIX:
-        assert report.task.valid
-
-
-def test_validate_stats_hash_sha256_invalid():
-    hash = "a1fd6c5ff3494f697874deeb07f69f8667e903dd94a7bc062dd57550cea26da8"
-    report = validate("data/table.csv", hashing="sha256", stats={"hash": "bad"})
-    if IS_UNIX:
-        assert report.flatten(["code", "note"]) == [
-            [
-                "hash-count-error",
-                'expected sha256 is "bad" and actual is "%s"' % hash,
-            ],
-        ]
-
-
-def test_validate_stats_hash_sha512():
-    hash = "d52e3f5f5693894282f023b9985967007d7984292e9abd29dca64454500f27fa45b980132d7b496bc84d336af33aeba6caf7730ec1075d6418d74fb8260de4fd"
-    report = validate("data/table.csv", hashing="sha512", stats={"hash": hash})
-    if IS_UNIX:
-        assert report.task.valid
-
-
-def test_validate_stats_hash_sha512_invalid():
-    hash = "d52e3f5f5693894282f023b9985967007d7984292e9abd29dca64454500f27fa45b980132d7b496bc84d336af33aeba6caf7730ec1075d6418d74fb8260de4fd"
-    report = validate("data/table.csv", hashing="sha512", stats={"hash": "bad"})
-    if IS_UNIX:
-        assert report.flatten(["code", "note"]) == [
-            [
-                "hash-count-error",
-                'expected sha512 is "bad" and actual is "%s"' % hash,
-            ],
-        ]
-
-
-def test_validate_stats_bytes():
-    report = validate("data/table.csv", stats={"bytes": 30})
-    if IS_UNIX:
-        assert report.task.valid
-
-
-def test_validate_stats_bytes_invalid():
-    report = validate("data/table.csv", stats={"bytes": 40})
-    assert report.task.error.get("rowPosition") is None
-    assert report.task.error.get("fieldPosition") is None
-    if IS_UNIX:
-        assert report.flatten(["code", "note"]) == [
-            ["byte-count-error", 'expected is "40" and actual is "30"'],
-        ]
-
-
-def test_validate_stats_rows():
-    report = validate("data/table.csv", stats={"rows": 2})
-    if IS_UNIX:
-        assert report.task.valid
-
-
-def test_validate_stats_rows_invalid():
-    report = validate("data/table.csv", stats={"rows": 3})
-    assert report.task.error.get("rowPosition") is None
-    assert report.task.error.get("fieldPosition") is None
-    if IS_UNIX:
-        assert report.flatten(["code", "note"]) == [
-            ["row-count-error", 'expected is "3" and actual is "2"'],
-        ]
-
-
-# Detector
-
-
-def test_validate_detector_sync_schema():
-    schema = {
-        "fields": [
-            {"name": "id", "type": "integer"},
-            {"name": "name", "type": "string"},
-        ],
-    }
-    detector = Detector(schema_sync=True)
-    report = validate("data/sync-schema.csv", schema=schema, detector=detector)
-    assert report.valid
-    assert report.task.resource.schema == {
-        "fields": [
-            {"name": "name", "type": "string"},
-            {"name": "id", "type": "integer"},
-        ],
-    }
-
-
-def test_validate_detector_sync_schema_invalid():
-    source = [["LastName", "FirstName", "Address"], ["Test", "Tester", "23 Avenue"]]
-    schema = {"fields": [{"name": "id"}, {"name": "FirstName"}, {"name": "LastName"}]}
-    detector = Detector(schema_sync=True)
-    report = validate(source, schema=schema, detector=detector)
-    assert report.valid
-
-
-def test_validate_detector_headers_errors():
-    source = [
-        ["id", "last_name", "first_name", "language"],
-        [1, "Alex", "John", "English"],
-        [2, "Peters", "John", "Afrikaans"],
-        [3, "Smith", "Paul", None],
-    ]
-    schema = {
-        "fields": [
-            {"name": "id", "type": "number"},
-            {"name": "language", "constraints": {"required": True}},
-            {"name": "country"},
-        ]
-    }
-    detector = Detector(schema_sync=True)
-    report = validate(source, schema=schema, detector=detector)
-    assert report.flatten(["rowPosition", "fieldPosition", "code", "cells"]) == [
-        [4, 4, "constraint-error", ["3", "Smith", "Paul", ""]],
-    ]
-
-
-def test_validate_detector_patch_schema():
-    detector = Detector(schema_patch={"missingValues": ["-"]})
-    report = validate("data/table.csv", detector=detector)
-    assert report.valid
-    assert report.task.resource.schema == {
-        "fields": [
-            {"name": "id", "type": "integer"},
-            {"name": "name", "type": "string"},
-        ],
-        "missingValues": ["-"],
-    }
-
-
-def test_validate_detector_patch_schema_fields():
-    detector = Detector(
-        schema_patch={"fields": {"id": {"type": "string"}}, "missingValues": ["-"]}
-    )
-    report = validate("data/table.csv", detector=detector)
-    assert report.valid
-    assert report.task.resource.schema == {
-        "fields": [{"name": "id", "type": "string"}, {"name": "name", "type": "string"}],
-        "missingValues": ["-"],
-    }
-
-
-def test_validate_detector_infer_type_string():
-    detector = Detector(field_type="string")
-    report = validate("data/table.csv", detector=detector)
-    assert report.valid
-    assert report.task.resource.schema == {
-        "fields": [{"name": "id", "type": "string"}, {"name": "name", "type": "string"}],
-    }
-
-
-def test_validate_detector_infer_type_any():
-    detector = Detector(field_type="any")
-    report = validate("data/table.csv", detector=detector)
-    assert report.valid
-    assert report.task.resource.schema == {
-        "fields": [{"name": "id", "type": "any"}, {"name": "name", "type": "any"}],
-    }
-
-
-def test_validate_detector_infer_names():
-    detector = Detector(field_names=["id", "name"])
-    report = validate(
-        "data/without-headers.csv",
-        layout={"header": False},
-        detector=detector,
-    )
-    assert report.task.resource.schema["fields"][0]["name"] == "id"
-    assert report.task.resource.schema["fields"][1]["name"] == "name"
-    assert report.task.resource.stats["rows"] == 3
-    assert report.task.resource.labels == []
-    assert report.task.resource.header == ["id", "name"]
-    assert report.valid
-
-
-# Validation
-
-
 def test_validate_pick_errors():
     report = validate("data/invalid.csv", pick_errors=["blank-label", "blank-row"])
     assert report.task.scope == ["blank-label", "blank-row"]
